OPA_APP/historic/requirements.txt
pandas
pip 
python-binance
rel
requests
websocket-client   
websockets
psycopg2-binary
